Create Done Tasks tiddlers.txt
List of all the completed actionable tiddlers. List is automatically populated when a //done// tag is added to tiddler. Tiddlers that appear in this list must have a //due// custom field, and a //done// tag. Items will be listed with a marked checkbox. If checkbox is unmarked from tiddler, the //done// tag is removed from that particular tiddler and its entry is removed from the Done list.